requirements.txt
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dj-database-url==2.1.0
psycopg2-binary==2.9.9
requests==2.31.0
fastapi
uvicorn[standard]
pydantic
pydantic-settings
torch
transformers
sentencepiece
drf-spectacular
python-dotenv
